split up calculate datamodel, sanitize all datamodel names after the complete datamodel was calculated
