+++ introspection-engine/connectors/sql-introspection-connector/tests/unit_tests/mod.rs
@@ -3,7 +3,7 @@
 mod get_metadata_command_for_empty_db_tests;
 mod get_metadata_command_tests;
 mod list_databases_command_tests;
-mod mapping_invalid_characters;
 mod mysql;
 mod postgres;
+mod remapping_invalid_characters;
 mod sqlite;
